Move the conversion from color index to color into the shader. Like this, only the index of the color is set on the CPU. The conversion to color is set on the GPU and the engine doesn't need to pass over the full image again.

+++ tiny-engine/src/commonMain/kotlin/com/github/minigdx/tiny/lua/SprLib.kt
@@ -49,7 +49,7 @@ class SprLib(val gameOptions: GameOptions, val resourceAccess: GameResourceAcces
             if (isInPixelArray(pixelArray, x, y)) {
                 val index = pixelArray.get(x, y)
                 val colorIndex = index.get(0)
-                return valueOf(colorIndex)
+                return valueOf(colorIndex.toInt())
             } else {
                 return NIL
             }
